+++ CHANGELOG.md
@@ -32,5 +32,6 @@
 ### Fixed
 
 -   `Field.select` label size and tag can be changed
--   `DateInputField` raises a `ValidationError` (instead of `ValueError`) when given invalid input
+-   `DateInputField` no longer raises a `ValueError` when given invalid input (a `ValidationError` is raised instead)
 -   `DateInputField` with `required=False` no longer raises a `ValueError` when no values are passed
+-   `DateInputField` no longer errors with `OverflowError` when large values are passed